refactor(auth): API 명세 변경에 따른 리프레시 로직 수정
- 토큰 리프레시 API의 엔드포인트가 `/auth/refresh`에서 `/token`으로 변경되었습니다.
- 기존에 Header로 전달하던 리프레시 토큰을 Request Body에 담아 보내도록 수정했습니다.
- 관련 데이터 모델(`LoginRequest`, `LoginResponse`)들을 `auth` 하위 패키지로 이동하고, `RefreshTokenRequest` 모델을 추가했습니다.
