+++ data/src/main/java/com/smtm/pickle/data/mapper/AuthMapper.kt
@@ -1,6 +1,6 @@
 package com.smtm.pickle.data.mapper
 
-import com.smtm.pickle.data.source.remote.model.LoginResponse
+import com.smtm.pickle.data.source.remote.model.auth.LoginResponse
 import com.smtm.pickle.domain.model.auth.AuthToken
 
 fun LoginResponse.toDomain(): AuthToken = AuthToken(
