+++ data/src/main/java/com/smtm/pickle/data/repository/AuthRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.smtm.pickle.data.repository
 import com.smtm.pickle.data.mapper.toDomain
 import com.smtm.pickle.data.source.remote.api.AuthService
 import com.smtm.pickle.data.source.remote.datasource.GoogleAuthDataSource
-import com.smtm.pickle.data.source.remote.model.LoginRequest
+import com.smtm.pickle.data.source.remote.model.auth.LoginRequest
 import com.smtm.pickle.domain.model.auth.AuthToken
 import com.smtm.pickle.domain.model.auth.SocialLoginType
 import com.smtm.pickle.domain.provider.TokenProvider
