+++ data/src/main/java/com/smtm/pickle/data/source/remote/api/AuthService.kt
@@ -1,7 +1,7 @@
 package com.smtm.pickle.data.source.remote.api
 
-import com.smtm.pickle.data.source.remote.model.LoginRequest
-import com.smtm.pickle.data.source.remote.model.LoginResponse
+import com.smtm.pickle.data.source.remote.model.auth.LoginRequest
+import com.smtm.pickle.data.source.remote.model.auth.LoginResponse
 import retrofit2.http.Body
 import retrofit2.http.POST
