+++ data/src/main/java/com/smtm/pickle/data/source/remote/api/RefreshTokenApi.kt
@@ -1,16 +1,17 @@
 package com.smtm.pickle.data.source.remote.api
 
-import com.smtm.pickle.data.source.remote.model.LoginResponse
-import retrofit2.http.Header
+import com.smtm.pickle.data.source.remote.model.auth.LoginResponse
+import com.smtm.pickle.data.source.remote.model.auth.RefreshTokenRequest
+import retrofit2.http.Body
 import retrofit2.http.POST
 
 /** Authenticator 내부에서만 사용할 API */
 interface RefreshTokenApi {
 
     /** Authenticator 내부 호출용 API */
-    @POST("auth/refresh")
+    @POST("token")
     suspend fun refreshToken(
-        @Header("Authorization") refreshToken: String
+        @Body refreshToken: RefreshTokenRequest
     ): LoginResponse
 
 }
