+++ data/src/main/java/com/smtm/pickle/data/source/remote/model/auth/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.smtm.pickle.data.source.remote.model
+package com.smtm.pickle.data.source.remote.model.auth
 
 import kotlinx.serialization.Serializable
