+++ data/src/main/java/com/smtm/pickle/data/source/remote/model/auth/RefreshTokenRequest.kt
@@ -1,0 +1,8 @@
+package com.smtm.pickle.data.source.remote.model.auth
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RefreshTokenRequest(
+    val refreshToken: String
+)
